# contracts/openapi/ops-api.yaml
openapi: 3.1.0
info:
  title: Dawnline ops-api
  description: |-
    운영 콘솔 API (DESIGN.md §5.5). 읽기 모델 조회와 운영자 커맨드다.

    **인증은 JWT 다.** `make token ROLE=…` 가 찍은 토큰을 `Authorization: Bearer` 로 싣는다 (만료 12시간). `GET` 은 `OPS_VIEWER`, 나머지는 `OPS_OPERATOR` 이상이다. 토큰이 없거나 유효하지 않으면 401 `unauthenticated`, 역할이 모자라면 403 `forbidden` 이다.

    **커맨드는 코어로 위임하고 감사한다.** 응답 헤더 `X-Dawnline-Audit-Id` 에 감사 행 id 가 온다. 코어가 거절하면(4xx) 코어의 상태와 Problem Details 본문을 그대로 돌려준다. 코어에 닿지 못하면 502 `core-unreachable`, 적용됐는지 모르면 504 `core-timeout` 또는 502 `core-error` 다 — 그때는 다시 누르기가 먼저다(RB-07).

    **조회는 감사하지 않는다.** 읽기 모델(`rm_*`)에서 읽고, 라우트의 stop 은 dispatch 에 조회를 위임한다 — 좌표와 순서의 진실은 dispatch 이고 읽기 모델은 집계다.

    오류 응답은 RFC 9457 Problem Details 이고 `type` 과 `code` 가 항상 채워진다.
  version: 1.0.0
servers:
- url: http://localhost:8085
  description: 로컬 개발
paths:
...
  /api/v1/routes/{routeId}/stops/{orderId}/reassign:
    post:
      tags:
      - ops-command-controller
      operationId: reassign
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: orderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignBody"
        required: true
      responses:
        "200":
          description: 두 라우트의 새 revision
          headers:
            X-Dawnline-Audit-Id:
              description: 감사 행 id — 모든 결과에 온다
              style: simple
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StopReassigned"
        "400":
          description: "`targetRouteId` 가 없다(감사 행 없음), 또는 코어의 거절 그대로 — 사유는 `dispatch-service.yaml`\
            \ 의 `reassign`"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: 코어의 거절 그대로 — 사유는 `dispatch-service.yaml` 의 `reassign`
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: 코어의 거절 그대로 — 사유는 `dispatch-service.yaml` 의 `reassign`
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "502":
          description: '`core-unreachable`(닿지 않았다 — 감사 `FAILED`) 또는 `core-error`(코어의
            5xx — 감사 `UNKNOWN`)'
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "503":
          description: 감사 행을 쓰지 못해 위임하지 않았다 — 기록 없는 커맨드는 없다
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "504":
          description: '`core-timeout` — 적용됐는지 모른다(감사 `UNKNOWN`). 다시 누르기가 먼저다(RB-07)'
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: 토큰이 없거나 유효하지 않다 — `unauthenticated`. 토큰을 다시 넣는다
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: "역할이 모자란다 — `forbidden`. 조회는 `OPS_VIEWER`, 커맨드는 `OPS_OPERATOR`\
            \ 이상"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - opsJwt: []
...
  /api/v1/routes/{routeId}:
    get:
      tags:
      - read-model-controller
      operationId: route
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: stop 순서·좌표·상태 — dispatch 에 조회를 위임한다(감사 없음)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "404":
          description: 코어의 거절 그대로 — 사유는 `dispatch-service.yaml` 의 `get`
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "502":
          description: '`core-unreachable` 또는 `core-error`'
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "504":
          description: '`core-timeout`'
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: 토큰이 없거나 유효하지 않다 — `unauthenticated`. 토큰을 다시 넣는다
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: "역할이 모자란다 — `forbidden`. 조회는 `OPS_VIEWER`, 커맨드는 `OPS_OPERATOR`\
            \ 이상"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - opsJwt: []
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          type: string
          description: "오류 코드 — `type` 의 마지막 세그먼트와 같다. 모든 오류에 실린다. 그 밖의 확장 멤버(`retryAfterSeconds`,\
            \ 검증 오류, 도메인 예외의 상세)도 이 칸처럼 최상위에 온다"
...
    ReassignBody:
      type: object
      properties:
        targetRouteId:
          type: string
          format: uuid
      required:
      - targetRouteId
    StopReassigned:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        fromRouteId:
          type: string
          format: uuid
        fromRevision:
          type: integer
          format: int32
        toRouteId:
          type: string
          format: uuid
        toRevision:
          type: integer
          format: int32
      required:
      - fromRevision
      - fromRouteId
      - orderId
      - toRevision
      - toRouteId
...
    RouteDetail:
      type: object
      properties:
        routeId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        status:
          type: string
        revision:
          type: integer
          format: int32
        distanceM:
          type: integer
          format: int32
        durationS:
          type: integer
          format: int32
        costKrw:
          type: integer
          format: int64
        stops:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
      required:
      - costKrw
      - distanceM
      - durationS
      - planId
      - revision
      - routeId
      - status
      - stops
      - vehicleId
    RouteStop:
      type: object
      properties:
        seq:
          type: integer
          format: int32
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        plannedArrival:
          type: string
          format: date-time
        status:
          type: string
        orderIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - lat
      - lng
      - orderIds
      - plannedArrival
      - seq
      - status
...
  securitySchemes:
    opsJwt:
      type: http
      description: '`make token ROLE=OPS_VIEWER|OPS_OPERATOR|ADMIN` 가 찍는 HS256 토큰
        (DESIGN.md §5.5)'
      scheme: bearer
      bearerFormat: JWT